ref: move zone component base to `Core`
