--- src/CoreLib/Game/Zone/Components/IServiceComponent.cs
@@ -1,30 +1,0 @@
-/* Copyright (C) Revive101 Development Team - All Rights Reserved
- * Unauthorized copying of this file, via any medium is strictly prohibited
- * Proprietary and confidential.
- */
-
-using Akka.Actor;
-using Imlight.CoreLib.WizardData.Models.Player;
-using System.Collections.Generic;
-using static Imlight.Common.Caches.TypeCache;
-
-namespace Imlight.CoreLib.Game.Zone.Components;
-
-/// <summary>
-/// Interface for service components that can be attached to a zone entity.
-/// </summary>
-public interface IServiceComponent {
-
-    IEnumerable<ServiceOptionBase> GetServiceOptions(Wizard playerCharacter);
-    void OnServiceInteraction(IActorRef playerActor, Wizard playerCharacter, CoreObject playerObject, uint serviceOptionIndex);
-    
-    string ServiceName { get; }
-    string NpcIcon { get; }
-    string NpcNameKey { get; }
-    string NpcTextKey { get; }
-    string WizBang { get; } 
-    string StateName { get; }
-    string InteractWizBang { get; }
-    string DisplayKey { get; }
-
-}
